rename parent package lifecycle to autobuild
